+++ src/pages/scripts/ydits-web/services/wolfx/data/jma-eew-warn-areas.js
@@ -1,0 +1,33 @@
+/**!
+ *
+ * YDITS for Web
+ *
+ * Copyright (C) よね/Yone
+ *
+ * Licensed under the Apache License 2.0.
+ *
+ */
+
+import { WolfxJmaEewWarnArea } from "./jma-eew-warn-area.js";
+
+/**
+ * Wolfx JMA EEW 警報地域のデータクラス
+*/
+export class WolfxJmaEewWarnAreas {
+    /**
+     * @type {WolfxJmaEewWarnArea[]}
+     */
+    areas = [];
+
+
+    /**
+     * @param {[Object<string, string|number|Date>]} areas - Wolfx JMA EEW 警報地域のJSONデータクラス
+     */
+    constructor(areas = []) {
+        areas.forEach(area => {
+            this.areas.push(
+                new WolfxJmaEewWarnArea(area)
+            );
+        });
+    }
+}
